method-1/requirements.txt
streamlit==1.39.0
ultralytics==8.3.30
opencv-contrib-python==4.10.0.84
scipy==1.13.1
librosa==0.10.2.post1
sounddevice==0.4.7
pywavelets==1.6.0
matplotlib==3.9.2
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
transformers==4.45.2
accelerate==1.0.1
einops==0.8.0
onnxruntime==1.19.2
Pillow==10.4.0
scikit-image==0.24.0
scikit-learn==1.5.2
soundfile==0.12.1
PyYAML==6.0.2
numpy>=1.26,<2.3

